requirements.txt
streamlit
fastapi
uvicorn
pillow
numpy
pandas
matplotlib
requests
torchvision
torch
ultralytics  # optional but recommended for YOLOv8 detection
